+++ Assets/_Project/Scripts/Core/Configuration/GameConfigAsset.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace PawsAndLoot.Config
+{
+    public abstract class GameConfigAsset : ScriptableObject
+    {
+        public abstract void ValidateOrThrow();
+    }
+}